requirements.txt
streamlit==1.39.0
openai>=1.52.0,<2.0.0
pydantic>=2.7.0,<3.0.0
python-dotenv>=1.0.0
Pillow>=10.0.0
